# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dominion Realm API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/settings/models": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Models",
        "description": "Every customizable agent's current model + tier (legacy endpoint).",
        "operationId": "get_models_settings_models_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettingsOut"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Set Model",
        "description": "Point one agent role at Haiku / Sonnet / Opus. Applies live + persists.",
        "operationId": "set_model_settings_models_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelSettingUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelSettingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/settings/presets/{preset_id}": {
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Apply Preset",
        "description": "Apply a built-in preset to all agent roles.",
        "operationId": "apply_preset_settings_presets__preset_id__put",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Preset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOpsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/settings/agents/{setting}/policy": {
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Set Agent Policy",
        "description": "Update fallback chain / never-fallback tiers for one agent.",
        "operationId": "set_agent_policy_settings_agents__setting__policy_put",
        "parameters": [
          {
            "name": "setting",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Setting"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentPolicyUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOpsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/settings/agents/smoke-test": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Smoke Test",
        "description": "Offline fixture smoke test \u2014 no API spend.",
        "operationId": "smoke_test_settings_agents_smoke_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/SmokeTestIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmokeTestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentContractOut": {
        "properties": {
          "inputs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Inputs"
          },
          "outputs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Outputs"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          },
          "max_retries": {
            "type": "integer",
            "title": "Max Retries",
            "default": 3
          },
          "context_load": {
            "type": "string",
            "title": "Context Load",
            "default": ""
          },
          "uses_memory": {
            "type": "boolean",
            "title": "Uses Memory",
            "default": false
          },
          "writes_artifacts": {
            "type": "boolean",
            "title": "Writes Artifacts",
            "default": false
          },
          "requires_approval": {
            "type": "boolean",
            "title": "Requires Approval",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "inputs",
          "outputs"
        ],
        "title": "AgentContractOut"
      },
      "AgentEstimateOut": {
        "properties": {
          "cost_band": {
            "type": "string",
            "title": "Cost Band"
          },
          "speed_band": {
            "type": "string",
            "title": "Speed Band"
          },
          "typical_calls_per_chapter": {
            "type": "integer",
            "title": "Typical Calls Per Chapter"
          }
        },
        "type": "object",
        "required": [
          "cost_band",
          "speed_band",
          "typical_calls_per_chapter"
        ],
        "title": "AgentEstimateOut"
      },
      "AgentOpsAgentOut": {
        "properties": {
          "setting": {
            "type": "string",
            "title": "Setting"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          },
          "policy": {
            "$ref": "#/components/schemas/AgentPolicyOut"
          },
          "contract": {
            "$ref": "#/components/schemas/AgentContractOut"
          },
          "permissions": {
            "$ref": "#/components/schemas/AgentPermissionsOut"
          },
          "estimate": {
            "$ref": "#/components/schemas/AgentEstimateOut"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "setting",
          "label",
          "description",
          "model",
          "policy",
          "contract",
          "permissions",
          "estimate"
        ],
        "title": "AgentOpsAgentOut"
      },
      "AgentOpsOut": {
        "properties": {
          "active_preset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Preset"
          },
          "presets": {
            "items": {
              "$ref": "#/components/schemas/AgentPresetOut"
            },
            "type": "array",
            "title": "Presets"
          },
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentOpsAgentOut"
            },
            "type": "array",
            "title": "Agents"
          },
          "pipeline_estimate": {
            "$ref": "#/components/schemas/PipelineEstimateOut"
          },
          "tiers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tiers"
          }
        },
        "type": "object",
        "required": [
          "active_preset",
          "presets",
          "agents",
          "pipeline_estimate",
          "tiers"
        ],
        "title": "AgentOpsOut"
      },
      "AgentPermissionsOut": {
        "properties": {
          "auto_run": {
            "type": "boolean",
            "title": "Auto Run",
            "default": true
          },
          "require_approval": {
            "type": "boolean",
            "title": "Require Approval",
            "default": false
          },
          "can_modify_packet": {
            "type": "boolean",
            "title": "Can Modify Packet",
            "default": false
          },
          "can_block_downstream": {
            "type": "boolean",
            "title": "Can Block Downstream",
            "default": false
          },
          "can_write_summaries": {
            "type": "boolean",
            "title": "Can Write Summaries",
            "default": false
          },
          "can_update_canon": {
            "type": "boolean",
            "title": "Can Update Canon",
            "default": false
          },
          "can_only_suggest": {
            "type": "boolean",
            "title": "Can Only Suggest",
            "default": true
          }
        },
        "type": "object",
        "title": "AgentPermissionsOut"
      },
      "AgentPolicyOut": {
        "properties": {
          "setting": {
            "type": "string",
            "title": "Setting"
          },
          "primary_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Tier"
          },
          "primary_model": {
            "type": "string",
            "title": "Primary Model"
          },
          "fallback_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Tier"
          },
          "fallback_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Model"
          },
          "never_fallback": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Never Fallback",
            "default": []
          },
          "escalation_rules": {
            "items": {
              "$ref": "#/components/schemas/EscalationRuleOut"
            },
            "type": "array",
            "title": "Escalation Rules",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "setting",
          "primary_model"
        ],
        "title": "AgentPolicyOut"
      },
      "AgentPolicyUpdateIn": {
        "properties": {
          "fallback_tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Tier"
          },
          "never_fallback": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Never Fallback"
          }
        },
        "type": "object",
        "title": "AgentPolicyUpdateIn"
      },
      "AgentPresetOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "cost_band": {
            "type": "string",
            "title": "Cost Band"
          },
          "latency_band": {
            "type": "string",
            "title": "Latency Band"
          },
          "best_for": {
            "type": "string",
            "title": "Best For"
          }
        },
        "type": "object",
        "required": [
          "id",
          "label",
          "description",
          "cost_band",
          "latency_band",
          "best_for"
        ],
        "title": "AgentPresetOut"
      },
...
      "EscalationRuleOut": {
        "properties": {
          "trigger": {
            "type": "string",
            "title": "Trigger"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "trigger",
          "description"
        ],
        "title": "EscalationRuleOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelSettingOut": {
        "properties": {
          "setting": {
            "type": "string",
            "title": "Setting"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          }
        },
        "type": "object",
        "required": [
          "setting",
          "label",
          "description",
          "model"
        ],
        "title": "ModelSettingOut",
        "description": "One customizable agent: its current model id + which tier (haiku/sonnet/opus) that is."
      },
      "ModelSettingUpdateIn": {
        "properties": {
          "setting": {
            "type": "string",
            "title": "Setting"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          }
        },
        "type": "object",
        "required": [
          "setting",
          "tier"
        ],
        "title": "ModelSettingUpdateIn",
        "description": "PUT body to point one agent role at a tier."
      },
      "ModelSettingsOut": {
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/ModelSettingOut"
            },
            "type": "array",
            "title": "Agents"
          },
          "tiers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tiers"
          }
        },
        "type": "object",
        "required": [
          "agents",
          "tiers"
        ],
        "title": "ModelSettingsOut"
      },
...
      "PipelineEstimateOut": {
        "properties": {
          "cost_band": {
            "type": "string",
            "title": "Cost Band"
          },
          "latency_band": {
            "type": "string",
            "title": "Latency Band"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "opus_calls": {
            "type": "integer",
            "title": "Opus Calls"
          },
          "sonnet_calls": {
            "type": "integer",
            "title": "Sonnet Calls"
          },
          "haiku_calls": {
            "type": "integer",
            "title": "Haiku Calls"
          },
          "total_estimated_calls": {
            "type": "integer",
            "title": "Total Estimated Calls"
          }
        },
        "type": "object",
        "required": [
          "cost_band",
          "latency_band",
          "summary",
          "opus_calls",
          "sonnet_calls",
          "haiku_calls",
          "total_estimated_calls"
        ],
        "title": "PipelineEstimateOut"
      },
...
      "SmokeTestAgentOut": {
        "properties": {
          "setting": {
            "type": "string",
            "title": "Setting"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/SmokeTestCheckOut"
            },
            "type": "array",
            "title": "Checks"
          }
        },
        "type": "object",
        "required": [
          "setting",
          "label",
          "passed",
          "checks"
        ],
        "title": "SmokeTestAgentOut"
      },
      "SmokeTestCheckOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "name",
          "ok"
        ],
        "title": "SmokeTestCheckOut"
      },
      "SmokeTestIn": {
        "properties": {
          "agents": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agents"
          }
        },
        "type": "object",
        "title": "SmokeTestIn"
      },
      "SmokeTestOut": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SmokeTestAgentOut"
            },
            "type": "array",
            "title": "Results"
          },
          "all_passed": {
            "type": "boolean",
            "title": "All Passed"
          }
        },
        "type": "object",
        "required": [
          "results",
          "all_passed"
        ],
        "title": "SmokeTestOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}